refactor(lunox-core): move common types to contracts
