+++ packages/lunox-core/src/Contracts/Exception/Handler.ts
@@ -1,5 +1,5 @@
 import type Response from "../../Http/Response";
-import type { Request } from "../../Http/Request";
+import type Request from "../../Http/Request";
 
 export interface ExceptionHandler {
   render(req: Request, e: Error): Promise<Response>;
