+++ packages/lunox-core/src/Contracts/Http/Middleware.ts
@@ -1,4 +1,4 @@
-import type { Request } from "../../Http/Request";
+import type Request from "../../Http/Request";
 import type { Middleware as PolkaMiddleware } from "polka";
 import type Response from "../../Http/Response";
