+++ packages/lunox-core/src/Contracts/Response.ts
@@ -1,5 +1,5 @@
 import type Response from "../Http/Response";
-import type { Request } from "../Http/Request";
+import type Request from "../Http/Request";
 
 export interface ResponseRenderer {
   render(request: Request): Promise<Response>;
