+++ packages/lunox-core/src/Contracts/index.ts
@@ -1,3 +1,5 @@
+import { Request } from "../Http";
+
 export * from "./Exception/Handler";
 export * from "./Exception/HttpExceptionInterface";
 export * from "./Foundation/Bootstrapper";
@@ -10,4 +12,18 @@ export * from "./Encryption";
 export * from "./Request";
 export * from "./Validation";
 export * from "./Response";
-export * from "../Types";
+
+export type {
+  Polka as Server,
+  Request as ServerRequest,
+  Response as ServerResponse,
+} from "polka";
+export type Concrete = NewableFunction | (() => any);
+
+export type CallBack = (...params: any[]) => any;
+export type Class<I, Args extends any[] = any[]> = new (...args: Args) => I;
+
+export type OnServer = <T = any>(
+  req: Request | undefined,
+  ctx: T,
+) => Promise<Record<string, any>>;
